+++ resources/image.rb
@@ -28,7 +28,7 @@ def initialize(*args)
 end
 
 attribute :image_url, kind_of: String
-attribute :image_type, kind_of: String, default: 'unknown', equal_to: ['unknown', 'ami', 'qcow']
+attribute :image_type, kind_of: String, default: 'unknown', equal_to: %w(unknown ami qcow vhd vmdk vdi iso raw)
 attribute :image_name, kind_of: String, default: 'default'
 attribute :identity_user, kind_of: String
 attribute :identity_pass, kind_of: String